requirements.txt
cassandra-driver==3.29.0
click==8.1.7
geomet==0.2.1.post1
py4j==0.10.9.7
pyspark==3.5.0
six==1.16.0
spark==0.2.1
